[master]保存领取规则和奖品规则

+++ nas-common/src/main/java/com/minyan/nascommon/po/RewardRulePO.java
@@ -16,10 +16,10 @@ public class RewardRulePO implements Serializable {
   private Long id;
 
   /** 活动id */
-  private Long activityId;
+  private Integer activityId;
 
   /** 模块id */
-  private Long moduleId;
+  private Integer moduleId;
 
   /** 事件id */
   private Long eventId;
